% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mmu_genes-data.R
\docType{data}
\name{Mmu_genes}
\alias{Mmu_genes}
\title{Mmu_genes}
\format{
An object of class \code{data.frame} with 144778 rows and 4 columns.
}
\source{
\href{https://www.ensembl.org/index.html}{ensembl biomaRt}
}
\usage{
Mmu_genes
}
\description{
A gene ID table for \emph{D. melanogaster}.
}
\details{
This table was built directly from the ensembl biomaRt.
}
\section{Format}{

A dataframe with the folowing columns :
\itemize{
\item \code{ens_id}: Ensembl ID (\emph{e.g} \code{ENSMUSG00000064367})
\item \code{transcript_id}: Transcript ID (\emph{e.g} \code{ENSMUST00000082418})
\item \code{public_name}: Gene name (\emph{e.g} \code{mt-Nd5})
\item \code{transcript_length}: Transcript length from 5'UTR start to 3'UTR end (\emph{e.g} \code{1824})
}
}

\keyword{datasets}
